namespace resource added
I added a working namespace resource and tests.

I chose to create a utils.IsNotFound() function because I think we'll be doing that a lot and it makes a big difference in how we handle the error. Not found in a read means remove it from state and in a delete means everything is kosher but is REALLY BAD after a create :joy:

I mocked the namespace client but haven't done anything interesting with it yet. I'd like to validate that my 404 stuff is working.

I added the models directory with a separate model struct because I felt it was helpful and that they might end up being used outside the package. Dunno if that will prove true but I like it thus far.

I tried to dump the state returns and read that but didn't find anything. While I'm following what terraform prescribes in their docs and what I see google doing, not being able to see the changes persisted in state after a test worries me.