successfully creating a namespace
ValueString() seems to be required to get the correct values from the models.

I added a check to the cloudv2 client to ensure we're actually getting an auth token. That was pretty irritating to work out before I got the debugging working without a solid error :sweat_smile:

I have absolutely validated that the create is in fact creating the namespace. I created one, then checked it in the dev--  ui.

I think I've done everything correctly for the destroy plan to work and testing it after the create test has resulted in no more create namespace. Reruns also don't create an error about it already existing on second test run.