improved namespace acceptance tests
I want testing to be driven by HCL in an easily viewed examples directory that can be sourced for documentation. That necessitated a minor rewrite to move some stuff around. I also got a better feel for what I wanted where.

Acceptance tests are moved to the tests/ directory due to import cycle issues when colocating them with their resource caused by the need for provider to be imported and the reality that provider imports resources.

After some fiddling I was able to get an examples directory to work as the resource for testing. This is good! My initial attempt used ProviderFactories defined at the Test level with the config passed in through heredocs mixed with environment variables. I wasn't able to define provider inside the examples files because TF threw an error that I was defining the provider at both the step and the test level, which meant everything had to be passed into Config.  But I worked out that if I declared providerfactories at the step level I could define provider it in the examples files without issue.

The import test isn't working yet and I'll need to hit it with a debugger until it does.

I am still not happy with the location of the redpanda provider and I still don't know where to put it. A provider directory is the obvious option, but then I have endless import renames and confusion caused by it so that's out. And I already called the main directory redpanda. Grr. Open to suggestions!