add network resource and testing
The network resource finally forced me to add pend functions in so I stuck those in the utils package with everything else weird. Acceptance testing is passing so we should be good to go with this one.

I still need to add support for timeout passthrough and I haven't tested state imports yet.