added cluster resource + acceptance tests
I added a cluster resource following the pattern of the other resources along with a first pass acceptance test that generates a GCP cluster.

The most important decision made here was the choice I made for how to handle redpanda_version and whether to persist the API returned value into state if the user does not provide one. Ultimately I opted to not persist the API returned value when the user value is nil. My rationale is preserved in a comment in the read function.

The unit test was added to catch more extraneous double quotes added by terraform which were causing the api to fail. Extremely irritating!

Allow deletion was added to ensure that user data cannot be accidentally deleted. Users must explicitly set allow_deletion to true before the request will be sent to the API. This is SOP for any provider that supports user data stores.