added aws test and parallel tests
Fixed an issue caused by another version of not found in the client's response. Really wish it came with bundled http error codes in the message/a status but it doesn't appear to.

Fixed a missing negation that caused the logic to work the opposite of how it was supposed to work.

I'm pretty sure that names are global unique -- or at least org unique -- rather than namespace which is pretty irritating and not something I saw in the RFC/PRD.