rejig auth and remove auth token
After some discussion I validated that the auth flow is always always client_id and client_secret -> request token -> spawn clients. Users should not be persisting the auth_token so I've removed it as an option.

I also added the ability to pass in client_id and client_secret through the environment which massively reduced the annoyance of working with the provider in testing.