added ns update testing and sweeping
In this commit I added update tests for namespace. To support this -- and reduce the annoyance of failed tests -- I also added a sweeper construct to clean up the results of failed tests.

Finally I added a quick function to get namespace by name. I didn't do much with pagination because my assumption is that we should find the correct name within the first 100 results given that we're providing the exact name to filter.

I also added a skip option for cluster tests because they are too slow for iterating through a new concept.