import test
Changed sweeper to just use the client since I needed it elsewhere anyway.

Also I apparently failed? to push the listing by name with the request somehow? Not sure how that happened, I could have sworn it was there and specifically called it out in the commit but I must have ctrl+z it somehow.

You can't!!!!! get namespaceId for import testing in the test case where the import is happening as an entry to the ImportStateId because it is called and verified immediately rather than waiting for its appropriate position in the run order for some unfathomable reason.