fix: add cluster state checking
Operations are failing without reporting an error or a status sometimes. It happens almost always on cluster creation but other long ops can make it happen. This is not in line with the API docs.

I have put in place a more reliable solution of checking the cluster state with polling at 1m intervals and a static 50m until I am able to get an API fix.