feat: rewrite AreWeDoneYet and add tests
AreWeDoneYet wasn't working the way I wanted. It was hard to read wasn't properly handling some situations. I've rewritten it to only care about completed or failed and to treat all errors the same (record but don't care) because many of them are intermittent or recoverable and we're more interested in the outcome of the state machine.

I also added a mock and test cases around the function to allow quicker iteration on what is surely one of the more critical components of the codebase.