fix: state inconsistency issue
ListClusters does not return most fields which was causing state inconsistency when using it to set the cluster url.

Using the GetTags return value is risky as the API adds a default managed by redpanda tag. We instead request our own tags and persist the ones we requested rather than the returned values to avoid the need for filtering default tags.