feat: make changes to support local testing
The new changes allow the user to build a provider binary and use it to run any of the cluster examples for testing purposes. It will generate a random name and then persist it (git knows to ignore the file) across make runs so you can teardown/run again without issues.

Note that you must set REDPANDA_CLIENT_ID and REDPANDA_CLIENT_SECRET as usual, and must also set REDPANDA_CLOUD_ENVIRONMENT (probably to ign, but this allows you to target prd as well).

The intent is to support development by making it easier to build and test against a specific cluster (no deletion) repeatedly without having to fiddle around with the acceptance tests. This will be most useful when developing against the dataplane endpoints but should still have value for cluster creation testing as it is easier to fiddle the HCL directly against a cluster and then explore what went wrong.