feat: test tier in create
Our previous solution wasn't workable. The issue is that configure is called after schema so we aren't able to create a client that can be passed through to the validator. And unfortunately the Schema function is called with an empty struct for request so we can't access it there either (the original gameplan). An alternative path was passing connections/clients into the resource through the provider's Resources function but this is called before the provider's Configure so it also can't access the relevant client/connection.

This commit resolves the issue by eliminating the validator and handling tier validation prior to the cluster create call. This should slightly speed things along but isn't the meaningful advantage over having the API invalidate us that the prior validator method (which runs at the plan stage and is massively faster but unfortunately working off static info) brought. There is a possible solution but it comes from the API: don't require ID/Secret to access the throughput tier endpoint. This seems out of scope at this time.