feat: cluster updates allow disabling features as well as enabling them
Cheap update-needed check by comparing an UpdateRequest generated from the
current state with an UpdateRequest generated from the new plan.

The nice part of this approach is that it will continue working even if
additional fields are added to the cluster model, such as PrivateLink
status or KafkaAPI seed brokers, since those aren't included in the
UpdateRequest.

It currently doesn't do per-field comparisons, and sets an update mask
that includes every field we set. This seems fine? Though may cause
an issue if any specific field is blocked behind a feature flag (as
cloud_provider_tags currently is, though we don't update that right
now).