chore: modify generateUpdateRequest to calculate diff between ClusterUpdates
We can't just check if the elements of state and plan are equal, because some
structs will eventually have some non-user-configurable state inside of them.

For example, kafka_api has a seed_brokers element that could exist in state,
but not in the plan.

These aren't actually in Terraform yet but they exist in the Cloud API, so
prepare for them being in Terraform.