initial pass at byovpc network
Includes the custom value/type needed to support the structure for byovpc network.

Probably need to refactor model to stick all of them into a separate package as this will quickly become unreadable if we need more such structures.

Quick explanation of the why:
if you want to convert to/from a struct in TPF for a single nested attribute you need to supply a valid customtype which it can use for the conversion process from user HCL to struct. Custom value is used to go from struct to state.

From a code perspective it's pretty tangled and there is a lot of interface to satisfy. From an end user perspective it is very much cleaner (hopefully).

I still need to think through how this will work when supporting GCP.