ci: fix automated release
Makes goreleaser verbose so we can see the errors and properly converts the gpg key from base64.

Also fixes gpg was attempting to ask for the passphrase with an interactive request. This method of invoking it via goreleaser's config file prevents that.

From the manpage:

If this command is used with --batch, --pinentry-mode has been set to loopback, and one of the passphrase options (--passphrase, --passphrase-fd, or
              --passphrase-file) is used, the supplied passphrase is used for the new key and the agent does not ask for it.