feat: adds support for byovpc network
BYOVPC resources all have a CustomerManagedResources type. Because this is a complex entity with a mix of computed and set resources it cannot be handled with a SingleNestedAttribute. It would be possible to handle it with a custom defined object but this would require supporting a large number of interfaces -- potentially subject to change -- with a lot of complex interactions.

As such I opted for the future proof option and used a types.Object + converted to/from the proto spec and model within the function calls. This is a pretty standard approach and doesn't require much mucking about with the deep lore of the Terraform Plugin Framework.