feat: add aws byovpc cluster support
This had the same issues as the network byovpc implementation but at a much larger scale.

An additional wrinkle was that some of the fields outside of the CMR object were using a struct rather than a types.Object. This was fine in their initial implementation when they were fairly simple but was now an issue as a number of computed elements needed to be added.

This necessitated a fairly major operaton to get everything up to snuff.

I also moved some things to new files because it was -- in my subjective opinion -- getting pretty hard to read and understand what was going on in the cluster.go file.

I also added a validator because if connection is set to public with a byovpc cluster the whole thing explodes and that seemed bad. It depended on the existance of the CMR object so it couldn't be a simple string validator.