feat: add byovpc example
I apologize in advance for what you're going to see in this commit.

So basically the terraform acceptance testing framework for the terraform plugin framework doesn't support the use of terraform modules. Yes, you read that right. In this context "doesn't support" means "will not delete any of the resources created by a module if the test fails". The module in question creates hundreds of resources so this isn't an ideal outcome.

As a result I split the module out to a separate file and imported the arns of the essential created resources. This isn't ideal and isn't how end users should do this. I'm not sure how to square the circle on this one.