feat: add acceptance test support sorta
Because of the compromises in the prior commit regarding the lack of support for terraform modules in the terraform plugin framework acceptance test framework this acceptance test has a substantial manual element that needs to be scripted in the makefile and integrated into the pipeline. It also needs to have the many arns output by the module passed into the example.

I haven't done this yet because it's a pretty substantial bit of bash. It's a TODO!

I also added some init functions to set defaults on global variables because there is usually but not always a right answer here. The issue is that if cloud rolls out a patch it won't make it to preprod for a week usually. So we need to be able to pivot between the two and doing it manually is inconvenient. This works but I absolutely wouldn't want to do it this way outside of an acceptance test.