fix: set cloud provider correctly
Cloud provider wasn't set in the model because the user doesn't provide it in the config. Maybe they could at one point but they sure can't now! This was never caught as it was never relevant before BYOVPC was added.

The correct handling is to pull it from the API call and then translate it before passing it along to generate model. Once generateModel correctly has the necessary data it is able to do its thing correctly.