feat: refactor of network
Goal here was multipart:

* split datasource model from resource model
* simplify model and proto generation
* reduce risk of future provider inconsistency

I chose to turn model into an object with methods to simplify logic within the actual terraform object. You just call the model -- which no longer needs to be passed around -- and it does whatever. I also felt like it was helpful to have a formal and obvious separation between code related to satisfying the framework and code related to building the objects relevant to Redpanda.

Splitting datasource from resource was needed as the two are different in important ways. This is most obvious with the desired timeout feature but can crop up in many other areas. Some of the most annoying sources of inconsistency errors has been because of this divide.

I wanted to simplify model/proto generation as I understand those paths much better now so I figured I could cut a good whack of cruft out of the process. I also wanted to move from constants to function calls because I wanted the ability to make logical changes in there if I needed to do so without having to do that in the calling function.